utworzenie metody returnVehicle(), modyfikacje klas powiązanych z stworzoną metodą

+++ biblioteka/include/model/ClientType.h
@@ -17,7 +17,7 @@ public:
 
     virtual std::string getClientType() const;
 
-    virtual const int getDiscount(int RentPrice) const;
+    virtual const double getDiscount(double RentPrice) const;
 
 };
